+++ src/core/index.ts
@@ -3,3 +3,4 @@ export * from "./fsm-process.ts";
 export * from "./fsm-state.ts";
 export * from "./fsm-state-config.ts";
 export * from "./fsm-state-descriptor.ts";
+export * from "./fsm-transitions.ts";